skypulse-bot/requirements.txt
aiogram==3.13.1                 # работа с Telegram API
python-dotenv==1.0.0             # загрузка переменных окружения

# База данных (выберите одну или обе)
asyncpg==0.29.0                  # драйвер PostgreSQL (если будете использовать)
aiosqlite==0.19.0                 # драйвер SQLite (для простоты, но не для продакшна)
SQLAlchemy==2.0.30                # ORM для работы с БД
alembic==1.13.1                   # миграции БД

# Валидация и модели данных
pydantic==2.7.1                   # валидация данных (aiogram уже использует её)
pydantic-settings==2.2.0          # удобная работа с настройками

# Для будущего веб-сервера (если понадобится API)
fastapi==0.111.0                   # современный веб-фреймворк
uvicorn[standard]==0.30.1          # ASGI-сервер
aiohttp==3.9.5                     # асинхронный HTTP-клиент/сервер (альтернатива)

# Для кэширования и очередей (если понадобится)
redis==5.0.4                       # работа с Redis

# Для логирования и мониторинга
loguru==0.7.2                      # красивое логирование (опционально)